backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
asyncpg>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dateutil>=2.8.2
email-validator>=2.0.0
httpx>=0.24.0
pillow>=10.0.0
feedparser>=6.0.0
openai>=1.0.0
anthropic>=0.3.0
beautifulsoup4>=4.12.0